+++ docs/app/components/chat/Chat.vue
@@ -4,7 +4,7 @@ import type { ToolUIPart, DynamicToolUIPart } from 'ai'
 import type { IconComponent } from '@bitrix24/b24ui-nuxt'
 import { DefaultChatTransport, isToolUIPart, isReasoningUIPart, isTextUIPart, getToolName } from 'ai'
 import { Chat } from '@ai-sdk/vue'
-import { isReasoningStreaming, isToolStreaming } from '@bitrix24/b24ui-nuxt/utils/ai'
+import { isPartStreaming, isToolStreaming } from '@bitrix24/b24ui-nuxt/utils/ai'
 import { useMemoize } from '@vueuse/core'
 import * as theme from '#build/b24ui'
 import ProseStreamPre from '../prose/PreStream.vue'
@@ -277,7 +277,7 @@ defineShortcuts({
             <B24ChatReasoning
               v-if="isReasoningUIPart(part)"
               :text="part.text"
-              :streaming="isReasoningStreaming(message, index, chat)"
+              :streaming="isPartStreaming(part)"
               :icon="AiStarsIcon"
               chevron="leading"
             >